+++ CHANGELOG.md
@@ -10,13 +10,14 @@ and this project ~~adheres to~~
 
 ### Changed
 
-- Added Output type to the VmiHandler
 - VmiHandler::finished() is renamed to VmiHandler::check_completion(),
   which now returns an Option&lt;Output&gt; instead of a bool
 
 ### Added
 
 - Implemented handling of PFN changes in the PageTableMonitor
+- Added Output type to the VmiHandler
+- vmi_core::os::OsModule + VmiOs::modules() to get the list of loaded modules
 
 ### Fixed